sales-portal-ars-allocation-service/requirements.txt
missingno
pandas
snowflake-snowpark-python
snowflake-connector-python
snowflake-connector-python[pandas]
fastapi
uvicorn
python-dotenv
python-dateutil
numpy
pydantic
psycopg2-binary
SQLAlchemy
requests
boto3
mypy
pandas-stubs
types-requests
botocore
pytz
starlette
cryptography